Subpicture fixes, AFD started and minor fixes
git-svn-id: https://svn.code.sf.net/p/deinterlace/code/trunk/DScaler5@3571 6e12ab15-cf2f-4ffe-bf36-45892a6a5b89

+++ Filters/Decoders/MpegVideo/DSUtil.cpp
@@ -1,5 +1,5 @@
 ///////////////////////////////////////////////////////////////////////////////
-// $Id: DSUtil.cpp,v 1.1 2004-02-06 12:17:16 adcockj Exp $
+// $Id: DSUtil.cpp,v 1.2 2004-04-14 16:31:34 adcockj Exp $
 ///////////////////////////////////////////////////////////////////////////////
 //
 //	Copyright (C) 2003 Gabest
@@ -27,6 +27,11 @@
 // CVS Log
 //
 // $Log: not supported by cvs2svn $
+// Revision 1.1  2004/02/06 12:17:16  adcockj
+// Major changes to the Libraries to remove ATL and replace with YACL
+// First draft of Mpeg2 video decoder filter
+// Broken DScalerFilter part converted to new library
+//
 ///////////////////////////////////////////////////////////////////////////////
 
 
@@ -89,7 +94,7 @@ bool ExtractBIH(const AM_MEDIA_TYPE* pmt, BITMAPINFOHEADER* bih)
 	return(false);
 }
 
-bool ExtractDim(const AM_MEDIA_TYPE* pmt, int& w, int& h, int& arx, int& ary)
+bool ExtractDim(const AM_MEDIA_TYPE* pmt, int& w, int& h, long& arx, long& ary)
 {
 	w = h = arx = ary = 0;
